chore(lint): ST1003: should not use MixedCaps in package name; packageGenerator should be packagegenerator
